requirements.txt
# NOTE:
# It is recommended to pin exact dependency versions (==) to ensure
# reproducibility, stability, and consistent behavior across environments.
# This prevents unexpected breaking changes from upstream libraries.

python-dotenv==1.0.1
faiss-cpu==1.7.4
sentence-transformers==2.6.1
transformers>=4.34.0,<4.39.0

fastapi==0.110.0
uvicorn[standard]==0.27.1

numpy==1.26.4
pydantic==2.6.4

openai==1.12.0
httpx==0.27.0

# torch is installed separately in Dockerfile (CPU-only)